refactor(multisig): extract message-signing into an optional capability interface
Message-signing was required of every multisig implementer, but not every protocol can
support it — Solana Squads has no EIP-1271 equivalent and a multisig PDA has no private key,
so it can only stub-throw. Move message-signing into optional IMultisigReadOnlyMessageSigning
/ IMultisigMessageSigning interfaces (mirroring IMultisigOwnerManagement): protocols that
support it implement them; those that don't simply omit them.

- New src/multisig/multisig-message-signing.js: IMultisigReadOnlyMessageSigning
  (getMessageProposal/getMessageProposals) and IMultisigMessageSigning extends it
  (proposeMessage/approveMessageProposal); owns the MultisigMessageProposal and
  MultisigSignature typedefs.
- Removed those methods + typedefs from the core IWalletAccountReadOnlyMultisig /
  IWalletAccountMultisig.
- Updated the multisig barrel and hand-maintained .d.ts.

verify(message, signature) is deliberately left out of scope: it lives on the base
IWalletAccountReadOnlySimple (every wallet account, not just multisig), so making it optional
is a broader change with a larger blast radius — tracked separately to keep this PR atomic.
