+++ src/multisig/wallet-account-multisig.js
@@ -23,7 +23,6 @@ import { NotImplementedError } from '../errors.js'
 /** @typedef {import('../wallet-account.js').KeyPair} KeyPair */
 
 /** @typedef {import('./wallet-account-read-only-multisig.js').MultisigProposal} MultisigProposal */
-/** @typedef {import('./wallet-account-read-only-multisig.js').MultisigMessageProposal} MultisigMessageProposal */
 
 /** @typedef {import('../errors.js').NoSuchElementError} NoSuchElementError */
 /** @typedef {import('../errors.js').ValueError} ValueError */
@@ -36,11 +35,6 @@ import { NotImplementedError } from '../errors.js'
 /**
  * @typedef {Object} MultisigAutoExecuteResult
  * @property {TransactionResult} [transaction] - If auto execute is set to true and the method call triggers the execution of the proposal, this field is set to the corresponding transaction's result (i.e., hash and fee).
- */
-
-/**
- * @typedef {Object} MultisigSignature
- * @property {string} signature - The caller's signature.
  */
 
 /** @interface */
@@ -93,29 +87,6 @@ export class IWalletAccountMultisig extends IWalletAccountReadOnlyMultisig {
    */
   async propose (tx, transactionOptions) {
     throw new NotImplementedError('propose(tx, transactionOptions)')
-  }
-
-  /**
-   * Proposes signing a message.
-   *
-   * @param {string} message - The message to sign.
-   * @returns {Promise<MultisigMessageProposal & MultisigSignature>} The multisig message proposal.
-   * @throws {Error} If the account is not an owner of the multisig wallet.
-   */
-  async proposeMessage (message) {
-    throw new NotImplementedError('proposeMessage(message)')
-  }
-
-  /**
-   * Approves an existing message proposal.
-   *
-   * @param {string} messageId - The message's hash.
-   * @returns {Promise<MultisigMessageProposal & MultisigSignature>} The multisig message proposal.
-   * @throws {Error} If the account is not an owner of the multisig wallet.
-   * @throws {NoSuchElementError} If no message exists for the given id.
-   */
-  async approveMessageProposal (messageId) {
-    throw new NotImplementedError('approveMessageProposal(messageId)')
   }
 
   /**
