+++ src/multisig/wallet-account-read-only-multisig.js
@@ -37,15 +37,6 @@ import { NotImplementedError } from '../errors.js'
  * @property {'pending' | 'executed'} status - The proposal's lifecycle state: `'pending'` while it still awaits confirmations or on-chain execution, `'executed'` once it has been executed on-chain.
  */
 
-/**
- * @typedef {Object} MultisigMessageProposal
- * @property {string} messageId - The message's hash.
- * @property {string} message - The original message.
- * @property {number} confirmations - The current number of confirmations.
- * @property {number} threshold -  The minimum amount of confirmations to sign the message.
- * @property {string | null} combinedSignature - The final combined signature when the threshold is met.
- */
-
 /** @interface */
 export class IWalletAccountReadOnlyMultisig extends IWalletAccountReadOnlySimple {
   /**
@@ -76,27 +67,6 @@ export class IWalletAccountReadOnlyMultisig extends IWalletAccountReadOnlySimple
    */
   async getProposal (proposalId) {
     throw new NotImplementedError('getProposal(proposalId)')
-  }
-
-  /**
-   * Returns a list of message proposals by their hashes.
-   *
-   * @param {string[]} messageIds - The list of message hashes
-   * @returns {Promise<Record<string, MultisigMessageProposal | null>>} For each message hash, the message details or
-   *   null if the message has not been found.
-   */
-  async getMessageProposals (messageIds) {
-    throw new NotImplementedError('getMessageProposals(messageIds)')
-  }
-
-  /**
-   * Returns a message proposal by its identifier.
-   *
-   * @param {string} messageId - The message's hash.
-   * @returns {Promise<MultisigMessageProposal | null>} The message details, or null if it has not been found.
-   */
-  async getMessageProposal (messageId) {
-    throw new NotImplementedError('getMessageProposal(messageId)')
   }
 
   /**
