+++ types/src/multisig/wallet-account-read-only-multisig.d.ts
@@ -22,21 +22,6 @@ export interface IWalletAccountReadOnlyMultisig extends IWalletAccountReadOnlySi
      */
     getProposal(proposalId: string): Promise<MultisigProposal | null>;
     /**
-     * Returns a list of message proposals by their hashes.
-     *
-     * @param {string[]} messageIds - The list of message hashes
-     * @returns {Promise<Record<string, MultisigMessageProposal | null>>} For each message hash, the message details or
-     *   null if the message has not been found.
-     */
-    getMessageProposals(messageIds: string[]): Promise<Record<string, MultisigMessageProposal | null>>;
-    /**
-     * Returns a message proposal by its identifier.
-     *
-     * @param {string} messageId - The message's hash.
-     * @returns {Promise<MultisigMessageProposal | null>} The message details, or null if it has not been found.
-     */
-    getMessageProposal(messageId: string): Promise<MultisigMessageProposal | null>;
-    /**
      * Quotes the on-chain cost of executing a pending proposal.
      *
      * @param {string} proposalId - The proposal's id.
@@ -76,28 +61,6 @@ export type MultisigProposal = {
      * - The proposal's lifecycle state: `'pending'` while it still awaits confirmations or on-chain execution, `'executed'` once it has been executed on-chain.
      */
     status: 'pending' | 'executed';
-};
-export type MultisigMessageProposal = {
-    /**
-     * - The message's hash.
-     */
-    messageId: string;
-    /**
-     * - The original message.
-     */
-    message: string;
-    /**
-     * - The current number of confirmations.
-     */
-    confirmations: number;
-    /**
-     * -  The minimum amount of confirmations to sign the message.
-     */
-    threshold: number;
-    /**
-     * - The final combined signature when the threshold is met.
-     */
-    combinedSignature: string | null;
 };
 export type TransactionResult = import("../wallet-account-read-only.js").TransactionResult;
 import { IWalletAccountReadOnlySimple } from '../wallet-account-read-only-simple.js';
